move area stats page problems by status calc into database
Doing this calculation in code turns out to be much too slow.

As part of this also fix an issue where if a report changed state last
month but had a further comment this month that was counted as being
a state change this month.

Also tweak the 'last month' start date to be midnight so the stats don't
change throughout the day.